# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Elaniin",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/users/": {
            "get": {
                "tags": [
                    "users End points"
                ],
                "summary": "Get the list of users",
                "operationId": "da52e0652a921c82fdc0b47e089dab67",
                "responses": {
                    "200": {
                        "description": "Return a list of users"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/11": {
            "get": {
                "tags": [
                    "users End points"
                ],
                "summary": "Get One user by id",
                "operationId": "66f10addeea12ecbe9120962331bce3d",
                "responses": {
                    "200": {
                        "description": "Return one user"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/search": {
            "post": {
                "tags": [
                    "users End points"
                ],
                "summary": "search user by User Name, Name and Email",
                "operationId": "2b9592a93af39d4722605952164c926d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "example": "admin"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "example": "admin@admin.com"
                    },
                    {
                        "name": "userName",
                        "in": "query",
                        "description": "UserName"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return one user"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/store": {
            "post": {
                "tags": [
                    "users End points"
                ],
                "summary": "Save user",
                "operationId": "961251cffc8049377e6b206f61f66fc1",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "example": "admin"
                    },
                    {
                        "name": "userName",
                        "in": "query",
                        "description": "UserName",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "example": "test@gmail.com"
                    },
                    {
                        "name": "birthDate",
                        "in": "query",
                        "description": "birthDate",
                        "example": "2062-02-26"
                    },
                    {
                        "name": "telephone",
                        "in": "query",
                        "description": "telephone",
                        "example": "77553344"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "example": "12345678"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "password_confirmation",
                        "required": true,
                        "example": "12345678"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data save successfully"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Oops something went wrong."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/update/9": {
            "put": {
                "tags": [
                    "users End points"
                ],
                "summary": "Update user",
                "operationId": "9efc7158b6da09d43c950fac49d0cdbf",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true
                    },
                    {
                        "name": "birthDate",
                        "in": "query",
                        "description": "birthDate",
                        "required": true
                    },
                    {
                        "name": "telephone",
                        "in": "query",
                        "description": "telephone",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data modify successfully"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Oops something went wrong."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users/7": {
            "delete": {
                "tags": [
                    "users End points"
                ],
                "summary": "Soft Delete One user by id",
                "operationId": "ba13470feb7a1d47f8f850b9d59fda41",
                "responses": {
                    "204": {
                        "description": "Data deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "500": {
                        "description": "Ups, fail request."
                    },
                    "default": {
                        "description": "Authorization Token not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}